feat(rpc): fetch manager index using its address

+++ codec/src/main/kotlin/acurast/codec/type/ProcessorOverview.kt
@@ -3,7 +3,7 @@ package acurast.codec.type
 import java.math.BigInteger
 
 public data class ProcessorOverview(
-    val managerIdentifier: Int? = null,
+    val managerIdentifier: BigInteger? = null,
     val isAttested: Boolean = false,
     val lastHeartbeatTimestamp: BigInteger? = null,
 )
